refactor(usb): centralize device identity resolution
